Reduce the contention on tmLock
There are mainly 4 functions will traverse shmGxactArray concurrently:
createDtx, releaseGxact, CreateDistributedSnapshot and getDtxCheckPointInfo

For CreateDistributedSnapshot, we use the same idea as GetSnapshotData in
upstream: strict serialization of commits and rollbacks with snapshot-taking.
So we acquire tmLock in shared mode in CreateDistributedSnapshot, and acquire
tmLock in exclusive mode when a proc exits from a transaction(clearTransactionState()).

For getDtxCheckPointInfo, we introduce a new flag to represent a global transaction has
committed on master, but hasn't finished the second phase of 2PC. This kind of global
transactions should be recorded in the checkpoint record. A new lwlock(CheckpointTMLock)
is used to protect the flag.