Fix the partition related panic that happens occasionally.
In function ATPExecPartSplit(rel, ...), heap_close and
then heap_open are performed several times on the relation
object referenced by rel.

The problem is, after heap_close call, the relcache entry
might be freed as its refcount goes to zero, and the following
heap_open call would return a different relcache entry pointer.
But in the caller of ATPExecPartSplit(), the old pointer is still
used to access the already-released relcache entry.

Since on most cases relcache entries would not be freed as soon as
their refcount goes to zero, this problem would not arise so often.
But in some parallel tests, especially with -DRELCACHE_FORCE_RELEASE
compile flag, this problem is expected to arise frequently.

This is a fix to alreay existing bug and current partition tests
can cover the code changes here. So there is no need to add new
test case.