Remove the restriction of not pulling up ANY sublink in the case of SRF.
Previously GPDB would refuse to pull up ANY sublink if the subquery
returns a set-returning function (SRF) in the targetlist, while
PostgreSQL will do that kind of pull-up.

This restriction was added in GPDB years ago to workaround MPP-7085.
That issue has been fixed in other place, so we can remove the
restriction now.