Fix qual pushdown through Set Operation with window functions.
Previously when checking if it is safe to push down a qual through
window functions in qual_is_pushdown_safe(), we did not consider the
case with Set Operations. As a result, qual would be incorrectly pushed
down through Set Operations with window functions.

In this patch, we recurse through setOperations tree when check for
window functions.

This patch fixes #5762.

Co-authored-by: Melanie Plageman <mplageman@pivotal.io>