Remove MERGE_FIXME in adjust_selectivity_for_nulltest.
Previously we have assertions in adjust_selectivity_for_nulltest to
ensure that the pushed qual of form "Var IS (NOT) NULL" is applied on
the nullable side of the outer join.

Now that we do not have the outer_rel and inner_rel passed in as
parameters, we cannot do that assertion any more. But we can know the
assertion is true, because otherwise the qual would be figured out as
not outerjoin_delayed by check_outerjoin_delay() and be pushed down to
the non-nullable side of the outer join.