Do not transform ShareInputScan in correlated subplan.
GPDB will transform the scan node in a correlated subplan to a tree looks
like:

Result (with quals)
       \
        \_Material
           \
            \_Broadcast (or Gather)
               \
                \_Scan (no quals)

Then the transformed plan can be executed in a parallel setting since the
correlation is now part of the result node which executes in the same slice
as the outer plan node. This works well for SeqScan.

For ShareInputScan, there would be one producer that have a SeqScan beneath
it, and the correlation originally exists with that SeqScan. If we do this
kind of transformation for ShareInputScan, the correlation is still with
that SeqScan.  And above this SeqScan, now there is a Broadcast motion
added. As we cannot pass param through motion, the plan would give wrong
results.

This patch avoid doing this transformation for ShareInputScan. Thus the
SeqScan beneath it would be transformed as above, and the correlation would
be pulled up to above the Broadcast motion.