Retain origianl RelOptInfos in two/three stage aggregate.
When making two/three stage aggregate, GPDB needs to rebuild arrays for
RelOptInfo and RangeTblEntry for the PlannerInfo when the underlying
range tables are transformed. Previously GPDB would just allocate a new
memory area for the arrays, without retaining the original RelOptInfos
and RangeTblEntrys. But it proves the original RelOptInfos are needed
later when we do set_plan_references processing on a SubqueryScan, which
would retrieve the RelOptInfo corresponding to the SubqueryScan.