Find the correct RelOptInfo for subquery when processing SubqueryScan.
When making two/three stage aggregate, GPDB needs to rebuild arrays for
RelOptInfo and RangeTblEntry for the PlannerInfo when the underlying
range tables are transformed. To do that, rebuild_simple_rel_and_rte()
just allocates a new memory area for the arrays, without retaining the
original RelOptInfos. As a result, when processing SubqueryScan for
subquery formed in set_base_rel_sizes(), we cannot find its RelOptInfo.

This patch tries to resolve this problem by recording the origin
RelOptInfo for subquery formed in set_base_rel_sizes() in
plan->subqueryRel.