Make Multi-DQA on subquery work.
When generating plan for multi-DQA, we need to build a coplan on base of
the shared scan for each distinct DQA, and then join the coplans back
together. And while building the coplan for each DQA, the arrays for
RelOptInfo and RangeTblEntry for the PlannerInfo would be rebuild and
the original arrays would have been replaced. So, for the DQA other than
the first one, we need to restore the arrays to what they are like for
the original shared scan plan before building coplan for it.