Fix up enable_hashagg when creating a UniquePath.
In PostgreSQL, semi-joins are implemented with JOIN_SEMI join types, or
by first eliminating duplicates from the inner side, and then performing
normal inner join (that's JOIN_UNIQUE_OUTER and JOIN_UNIQUE_INNER).

GPDB has a third way to implement them: Perform an inner join first, and
then eliminate duplicates from the result. This is performed by a
UniquePath above the join. And there are two ways to implement the
UniquePath: sorting and hashing. But if hashing is not enabled, we
should not consider it.

This patch fixes github issue #8437.