Fix how must_gather is determined with LIMIT ALL.
If there is ORDER BY or DISTINCT in the query, we need to bring all the
data to a single node by setting must_gather to be true. An exception is
when there's a LIMIT or OFFSET clause, which would be handled later when
inserting Limit node. Here to tell if there is any LIMIT or OFFSET
clause, we should use limit_needed, instead of checking limitCount or
limitOffset directly.

Fixes issue #9746.