Fix parameterized paths
This patch fixes two issues related to parameterized path logic on
master.

1. When generating unique row ID on the outer/inner side for join
JOIN_DEDUP_SEMI/JOIN_DEDUP_SEMI_REVERSE, we need to pass the param info
of outerpath/innerpath to the projection path. Otherwise we would have
problems when deciding whether a joinclause is movable to this join rel.

2. We should not pick up the parameterized path when its required outer
is beyond a Motion, since we cannot pass a param through Motion.

Fixes issue #10012