Fix wrong results on multiphase aggregation with ORDER BY
In the multiphase aggregation logic, cdb_grouping_planner would change
the parsetree, especially the range table list and targetlist. But for
EquivalenceClasses, their expressions are not changed accordingly, and
they are still based on the original range table list. As a result, when
we apply ORDER BY above the multiphase aggregation, in high possibility
we would be using the wrong EC for pathkey.

This PR clones a new root and resets its EC list when re-writing
sort_pathkeys, which is very hacky but works for this case.