Fix wrong results caused by over-eager constraint exclusion
When checking relation exclusion by constraints, GPDB performs further
checks that if the predicate is of format expr=constant, try to evaluate
the clause by replacing every occurrence of expr with the constant. This
works in some cases but not for all cases.

This PR fixes this issue #10287 by constraining the input types of
predicate to be types that we know we're safe with. For now we only use
integer.

This PR also fixes issue #8582 by checking if the restrictinfo_list
contains any nonstrict construct, so it can behavior correctly for NULL
tuples.