Fix wrong results for nullable constraint
When checking relation exclusion by constraints, GPDB performs further
checks that if the predicate is of format expr=constant, try to evaluate
the clause by replacing every occurrence of expr with the constant. This
works in some cases but not for all cases.

One such case is when the clause contains any nonstrict construct. In
this case we should not evaluate the clause with the constant in the
predicate. Otherwise we will lose NULL tuples.

Fixes issue issue #8582