Fix crash when creating partitioned table in utility mode
When transforming DISTRIBUTED BY clause in utility mode, NULL would be returned
as only QD can have policies. As a result, the child table in a partitioned
table would have NULL distributedBy, which would trigger assertion failure or
cause crash with current codes.

This patch fixes that by emitting an error message in that case to avoid the
crash.