Fix crash when planning for sub-queries that have grouping sets
When constructing plans for a list of rollups, the rollup_subplan may be
retrieved from root->simple_rel_array[i]->subplan, which is not always type of
SubqueryScan. So later in function rebuild_append_simple_rel_and_rte(), we need
to verify if the subplan is a SubqueryScan node before assigning a scanrelid to
it.

Fixes issue #10813 and issue 10841