Remove GPDB_12_MERGE_FIXME in create_unique_plan()
In GPDB, it's possible that the Unique plan node created in
create_unique_plan() is the topmost plan node, and the topmost plan node's
targetlist is required to be consistent with the original targetlist, to avoid
the assertion failure in apply_tlist_labeling(). So we need to change its
targetlist back if it has been changed.

In PostgreSQL, we don't have this same problem, because there would always be a
join node above the Unique plan node created here, which is projection capable.