Fix GroupingFunc for Plan nodes with SINGLETON flow
When creating Agg/Group nodes for rollups, previously we would skip the
final Agg for Plan nodes with Flow of FLOW_SINGLETON. This works fine in
most cases but not for GroupingFunc. As a result, queries such as the
one below would have wrong results for GroupingFunc.

  create table t1(a int, b int) distributed replicated;
  insert into t1 values(1,1);

  select grouping(a, b), a, b from t1 group by grouping sets((a, b),(a));
   grouping | a | b
  ----------+---+---
          0 | 1 | 1
          0 | 1 |
  (2 rows)

As a workaround, this PR just removes that logic and creates final Agg
if needed even for Plan nodes with Flow of FLOW_SINGLETON.