Improve NOT IN optimization with non-nullable OUTER and INNER
When performing LASJ to implement NOT IN, we need to check if the OUTER and
INNER are nullable. Once we can prove both OUTER and INNER are non-nullable, we
can skip broadcasting data from inner side, and we can also be able to choose
hash join even if there are multiple join keys.

Currently when there are multiple join keys, we do the nullable check
incorrectly for OUTER. This patch fixes that.