[POC] Eager aggregation for Greenplum
Eager aggregation is a kind of query transformation that partially
pushes a group-by past a join, so as to reduce the number of input rows
to the join and thus may result in a better overall plan. Currently
eager aggregation is impossible in Greenplum, because we always process
the scan/join portion first.

Eager aggregation can be illustrated in the plan below:

 EXPLAIN (COSTS OFF)
 SELECT a.i, avg(b.y + c.z)
 FROM a JOIN b ON a.i = b.j
     JOIN c ON c.k = a.i
 GROUP BY a.i;

 Gather Motion 3:1  (slice1; segments: 3)
   ->  Finalize GroupAggregate
         Group Key: a.i
         ->  Sort
               Sort Key: a.i
               ->  Hash Join
                     Hash Cond: (a.i = b.j)
                     ->  Seq Scan on a
                     ->  Hash
                           ->  Partial HashAggregate
                                 Group Key: b.j
                                 ->  Hash Join
                                       Hash Cond: (b.j = c.k)
                                       ->  Seq Scan on b
                                       ->  Hash
                                             ->  Seq Scan on c

This PR is just for fun and discussions. So consider it as a POC.

This PR is ported from PostgreSQL thread [1]. All credits belong there.
I made some changes to make it apply for Greenplum.

[1] https://www.postgresql.org/message-id/flat/9666.1491295317%40localhost